Creacion de salas privadas correcta
