+++ Camaleones/Assets/Scripts/Online/Debuging.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f23e44797d1bf54c869a25ed971d9aa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
