+++ Camaleones/Assets/Scripts/Online/Lobby/OnlineLobbyManager.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Photon.Pun;
+﻿using Photon.Pun;
 using Photon.Realtime;
 using TMPro;
 using UnityEngine;
